+++ src/kaleidoscope/plugin/LEDEffect-Rainbow.h
@@ -55,13 +55,13 @@ class LEDRainbowEffect : public Plugin,
     uint16_t rainbow_hue = 0;   //  stores 0 to 614
 
     uint8_t rainbow_steps = 1;  //  number of hues we skip in a 360 range per update
-    uint16_t rainbow_last_update = 0;
+    uint8_t rainbow_last_update = 0;
 
     byte rainbow_saturation = 255;
   };
 
  private:
-  uint16_t rainbow_update_delay = 40; // delay between updates (ms)
+  uint8_t rainbow_update_delay = 40; // delay between updates (ms)
   byte rainbow_value = 50;
 };
 
@@ -100,12 +100,12 @@ class LEDRainbowWaveEffect : public Plugin, public LEDModeInterface {
     uint16_t rainbow_hue = 0;  //  stores 0 to 614
 
     uint8_t rainbow_wave_steps = 1;  //  number of hues we skip in a 360 range per update
-    uint16_t rainbow_last_update = 0;
+    uint8_t rainbow_last_update = 0;
 
     byte rainbow_saturation = 255;
   };
 
-  uint16_t rainbow_update_delay = 40; // delay between updates (ms)
+  uint8_t rainbow_update_delay = 40; // delay between updates (ms)
   byte rainbow_value = 50;
 };
 }
